Created HttpParser and moved methods often used in previous commits
