+++ appserver/security/core-ee/src/main/java/com/sun/enterprise/security/ee/authentication/glassfish/jdbc/JDBCRealm.java
@@ -53,6 +53,7 @@ import javax.security.auth.login.LoginException;
 import javax.sql.DataSource;
 
 import org.glassfish.api.naming.SimpleJndiName;
+import org.glassfish.common.util.HttpParser;
 import org.glassfish.hk2.api.ActiveDescriptor;
 import org.glassfish.hk2.utilities.BuilderHelper;
 import org.jvnet.hk2.annotations.Service;
@@ -412,7 +413,7 @@ public final class JDBCRealm extends DigestRealmBase {
     private char[] hashPassword(char[] password) throws CharacterCodingException {
         byte[] bytes = null;
         char[] result = null;
-        Charset charSet = Utility.getCharset(getProperty(PARAM_CHARSET));
+        Charset charSet = HttpParser.getCharset(getProperty(PARAM_CHARSET));
         bytes = Utility.convertCharArrayToByteArray(password, charSet);
 
         if (messageDigest != null) {
