+++ appserver/tests/application/src/test/java/org/glassfish/main/test/app/mrjar/webapp/MultiReleaseWebApplicationTest.java
@@ -16,13 +16,12 @@
 
 package org.glassfish.main.test.app.mrjar.webapp;
 
-import com.sun.enterprise.util.Utility;
-
 import java.io.File;
 import java.io.IOException;
 import java.net.HttpURLConnection;
 import java.nio.file.Files;
 
+import org.glassfish.common.util.HttpParser;
 import org.glassfish.main.itest.tools.GlassFishTestEnvironment;
 import org.glassfish.main.test.app.mrjar.MultiReleaseTestBase;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
@@ -83,7 +82,7 @@ public class MultiReleaseWebApplicationTest extends MultiReleaseTestBase {
             assertAll(
                 () -> assertThat(connection.getResponseCode(), equalTo(200)),
                 // Check version of loaded class file
-                () -> assertThat(Integer.parseInt(Utility.readResponseInputStream(connection)), equalTo(V11))
+                () -> assertThat(Integer.parseInt(HttpParser.readResponseInputStream(connection)), equalTo(V11))
             );
         } finally {
             connection.disconnect();
@@ -100,7 +99,7 @@ public class MultiReleaseWebApplicationTest extends MultiReleaseTestBase {
             assertAll(
                 () -> assertThat(connection.getResponseCode(), equalTo(200)),
                 // Check version of loaded class file
-                () -> assertThat(Integer.parseInt(Utility.readResponseInputStream(connection)), equalTo(V17))
+                () -> assertThat(Integer.parseInt(HttpParser.readResponseInputStream(connection)), equalTo(V17))
             );
         } finally {
             connection.disconnect();
