+++ appserver/tests/application/src/test/java/org/glassfish/main/test/app/security/jmac/http/servlet/challenge/HttpServletChallengeAuthTest.java
@@ -28,9 +28,9 @@ import java.net.URLEncoder;
 import java.nio.file.Path;
 import java.util.Base64;
 
+import org.glassfish.common.util.HttpParser;
 import org.glassfish.main.itest.tools.GlassFishTestEnvironment;
 import org.glassfish.main.itest.tools.asadmin.Asadmin;
-import org.glassfish.main.test.app.security.common.HttpHeaderParser;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
 import org.jboss.shrinkwrap.api.exporter.ZipExporter;
 import org.jboss.shrinkwrap.api.spec.JavaArchive;
@@ -83,7 +83,7 @@ public class HttpServletChallengeAuthTest {
         createFileUser(FILE_REALM_NAME, USER_NAME2, USER_PASSWORD2, "mygroup");
 
         final JavaArchive loginModule = ShrinkWrap.create(JavaArchive.class)
-            .addClass(HttpServletChallengeTestAuthModule.class).addClass(HttpHeaderParser.class);
+            .addClass(HttpServletChallengeTestAuthModule.class).addClass(HttpParser.class);
         LOG.log(INFO, loginModule.toString(true));
         loginModuleFile = new File(getDomain1Directory().toAbsolutePath().resolve("../../lib").toFile(),
             AUTH_MODULE_NAME + ".jar");
