+++ nucleus/admin/util/src/main/java/com/sun/enterprise/admin/remote/sse/GfSseEventReceiverProprietaryReader.java
@@ -18,10 +18,11 @@
 package com.sun.enterprise.admin.remote.sse;
 
 import com.sun.enterprise.admin.remote.reader.ProprietaryReader;
-import com.sun.enterprise.util.Utility;
 
 import java.io.IOException;
 import java.io.InputStream;
+
+import org.glassfish.common.util.HttpParser;
 
 /**
  * Proprietary reader for SSE
@@ -35,6 +36,6 @@ public class GfSseEventReceiverProprietaryReader implements ProprietaryReader<Gf
 
     @Override
     public GfSseEventReceiver readFrom(final InputStream is, final String contentType) throws IOException {
-        return new GfSseEventReceiver(is, Utility.getCharsetFromHeader(contentType));
+        return new GfSseEventReceiver(is, HttpParser.getCharsetFromHeader(contentType));
     }
 }
