+++ nucleus/common/common-util/src/main/java/com/sun/enterprise/util/Utility.java
@@ -23,9 +23,7 @@ import java.lang.reflect.Array;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.lang.reflect.Modifier;
-import java.net.HttpURLConnection;
 import java.net.InetAddress;
-import java.net.URLConnection;
 import java.net.UnknownHostException;
 import java.nio.ByteBuffer;
 import java.nio.CharBuffer;
@@ -33,8 +31,6 @@ import java.nio.charset.CharacterCodingException;
 import java.nio.charset.Charset;
 import java.nio.charset.CharsetDecoder;
 import java.nio.charset.CharsetEncoder;
-import java.nio.charset.StandardCharsets;
-import java.nio.charset.UnsupportedCharsetException;
 import java.rmi.Remote;
 import java.security.AccessController;
 import java.security.PrivilegedAction;
@@ -50,8 +46,6 @@ import java.util.logging.Logger;
 import javax.naming.Context;
 import javax.naming.InitialContext;
 import javax.rmi.PortableRemoteObject;
-
-import static java.nio.charset.StandardCharsets.UTF_8;
 
 /**
  * Handy class full of static functions.
@@ -621,86 +615,6 @@ public final class Utility {
         }
 
         return envVal;
-    }
-
-    /**
-     * Reads response input stream from the connection to a String value.
-     * Respects charset from content type header, if not set, uses <code>UTF-8</code>.
-     *
-     * @param connection
-     * @return String parsed from {@link URLConnection#getInputStream()}
-     * @throws IOException
-     */
-    public static String readResponseInputStream(URLConnection connection) throws IOException {
-        Charset charset = getCharsetFromHeader(connection.getContentType());
-        try (InputStream in = connection.getInputStream()) {
-            return new String(in.readAllBytes(), charset);
-        }
-    }
-
-    /**
-     * Reads response error input stream from the connection to a String value.
-     * Respects charset from content type header, if not set, uses <code>UTF-8</code>.
-     *
-     * @param connection
-     * @return String parsed from {@link HttpURLConnection#getErrorStream()}
-     * @throws IOException
-     */
-    public static String readResponseErrorStream(HttpURLConnection connection) throws IOException {
-        Charset charset = getCharsetFromHeader(connection.getContentType());
-        try (InputStream in = connection.getErrorStream()) {
-            return new String(in.readAllBytes(), charset);
-        }
-    }
-
-    /**
-     * Parse the character encoding from the specified content type, authorization header
-     * or any header using same rules when they contain ie.
-     * <code>header="value"; charset=utf-8</code> .
-     * If the header is null, or there is no explicit character encoding,
-     * <code>UTF-8</code> is returned.
-     *
-     * @param header a content type header
-     * @return {@link Charset} or UTF-8 if the charset is null, empty string or unknown.
-     * @throws CharacterCodingException if the charset is not supported.
-     */
-    public static Charset getCharsetFromHeader(String header) throws CharacterCodingException {
-        if (header == null) {
-            return getCharset(null);
-        }
-        int start = header.indexOf("charset=");
-        if (start < 0) {
-            return getCharset(null);
-        }
-        String encoding = header.substring(start + 8);
-        int end = encoding.indexOf(';');
-        if (end >= 0) {
-            encoding = encoding.substring(0, end);
-        }
-        encoding = encoding.trim();
-        if (encoding.length() > 2 && encoding.startsWith("\"") && encoding.endsWith("\"")) {
-            encoding = encoding.substring(1, encoding.length() - 1);
-        }
-        return getCharset(encoding.trim());
-    }
-
-    /**
-     * Parses the parameter. If it is null or empty string, returns {@link StandardCharsets#UTF_8}.
-     * @param charset
-     * @return specified {@link Charset} or UTF-8 if the charset is null, empty string or unknown.
-     * @throws CharacterCodingException if the charset is not supported.
-     */
-    public static Charset getCharset(String charset) throws CharacterCodingException {
-        if (charset == null || charset.isEmpty()) {
-            return UTF_8;
-        }
-        try {
-            return Charset.forName(charset);
-        } catch (UnsupportedCharsetException uce) {
-            CharacterCodingException e = new CharacterCodingException();
-            e.initCause(uce);
-            throw e;
-        }
     }
 
     /**
